refactor: move the file into the services folder
